Fix compiler warnings
This commit makes changes to fix compiler warnings.


Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/95321
Tested-by: Chandrakala Chavva <cchavva@marvell.com>

+++ cmd/marvell/loki_sec_fw_update.c
@@ -13,7 +13,7 @@ static int do_sec_fw_update(struct cmd_tbl *cmdtp, int flag, int argc,
 			    char * const argv[])
 {
 	char *env1, *env2;
-	int addr, size;
+	unsigned long addr, size;
 	int ret;
 
 	if (argc == 1) {
